complete product routes test

+++ backend/Repositories/IProductRepository.cs
@@ -5,11 +5,11 @@ namespace Repositories
     public interface IProductRepository
     {
         Task<IEnumerable<Product>> GetAllProductsAsync();
-        Task<Product> GetProductByIdAsync(long id);
+        Task<Product> GetProductByIdAsync(int id);
         Task AddProductAsync(Product product);
         Task UpdateProductAsync(Product product);
-        Task DeleteProductAsync(long id);
-        Task<bool> ProductExistsAsync(long id);
+        Task DeleteProductAsync(int id);
+        Task<bool> ProductExistsAsync(int id);
         Task BulkAddProductsAsync(IEnumerable<Product> products);
     }
 }
